+++ timer.hpp
@@ -1,16 +1,17 @@
 #ifndef TIMER_HPP
 #define TIMER_HPP
 
+#include <cstdint>
 #include <ctime>
 
 class Timer {
 private:
-  int startedTime;
+  int64_t startedTime;
 public:
   Timer();
   ~Timer();
   void start();
-  int getTime();
+  int64_t getTime();
 };
 
 #endif
